Add test_info.py
